#3467: Rename tt_debug_print_server.hpp/cpp and move them out of llrt directory
