+++ tests/tt_metal/tt_metal/perf_microbenchmark/noc/test_noc_read_local_l1.cpp
@@ -12,7 +12,7 @@
 #include "test_tiles.hpp"
 #include "tt_metal/detail/tt_metal.hpp"
 #include "tt_metal/host_api.hpp"
-#include "tt_metal/llrt/tt_debug_print_server.hpp"
+#include "tt_metal/impl/debug/dprint_server.hpp"
 #include "tt_metal/test_utils/deprecated/tensor.hpp"
 #include "tt_metal/tools/profiler/op_profiler.hpp"
