+++ tt_metal/impl/module.mk
@@ -17,7 +17,8 @@ TT_METAL_IMPL_SRCS = \
 	tt_metal/impl/program/program.cpp \
 	tt_metal/impl/dispatch/device_command.cpp \
 	tt_metal/impl/dispatch/debug_tools.cpp \
-	tt_metal/impl/dispatch/command_queue.cpp
+	tt_metal/impl/dispatch/command_queue.cpp \
+	tt_metal/impl/debug/dprint_server.cpp
 
 TT_METAL_IMPL_OBJS = $(addprefix $(OBJDIR)/, $(TT_METAL_IMPL_SRCS:.cpp=.o))
 TT_METAL_IMPL_DEPS = $(addprefix $(OBJDIR)/, $(TT_METAL_IMPL_SRCS:.cpp=.d))
