+++ tt_metal/llrt/module.mk
@@ -7,7 +7,6 @@ LLRT_CFLAGS = $(CFLAGS) -Werror -Wno-int-to-pointer-cast
 
 LLRT_SRCS_RELATIVE = \
 	llrt/tt_cluster.cpp \
-	llrt/tt_debug_print_server.cpp \
 	llrt/llrt.cpp \
 	llrt/watcher.cpp \
 	llrt/rtoptions.cpp \
