+++ tt_metal/tools/profiler/tt_metal_profiler.cpp
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 #include "tt_metal/host_api.hpp"
-#include "llrt/tt_debug_print_server.hpp"
+#include "impl/debug/dprint_server.hpp"
 
 #include "tools/profiler/profiler.hpp"
